+++ covalent_results/app/core/config.py
@@ -35,7 +35,12 @@ class Settings(BaseSettings):
     RESULTS_SVC_HOST: str = "localhost"
     DATA_SVC_PORT: int = 8004
     DATA_SVC_HOST: str = "localhost"
-    RESULTS_DB: str = "results.db"
+    DATABASE_BACKEND: str = "sqlite"
+    DISPATCH_DB_NAME: str = "dispatch.db"
+    DISPATCH_DB_ENDPOINT: str = "localhost"
+    DISPATCH_DB_PORT: int = 3306
+    DISPATCH_DB_USER: str = "admin"
+    DISPATCH_DB_PASSWORD: str = "covalent"
 
     @validator("BACKEND_CORS_ORIGINS", pre=True)
     def assemble_cors_origins(cls, v: Union[str, List[str]]) -> Union[List[str], str]: